refactor: move search visibility modal open to separate func
